Claim: Handle claim phase rollout for single

+++ nunchuk-main/src/main/java/com/nunchuk/android/main/components/tabs/services/inheritanceplanning/claim/withdrawbitcoin/ClaimWithdrawBitcoinNavigation.kt
@@ -17,7 +17,7 @@ data object ClaimWithdrawBitcoinRoute
 
 fun NavGraphBuilder.claimWithdrawBitcoin(
     snackState: SnackbarHostState,
-    onNavigateToInputAmount: () -> Unit = {},
+    onNavigateToInputAmount: (Double) -> Unit = {},
     onNavigateToSelectWallet: () -> Unit = {},
     onNavigateToWalletIntermediary: () -> Unit = {},
     onNavigateToAddReceipt: () -> Unit = {},
